更新返回值类型和增加查询字段

+++ ord-api/src/main/java/rebue/ord/ro/OrdSettleRo.java
@@ -1,5 +1,7 @@
 package rebue.ord.ro;
 
+import java.math.BigDecimal;
+
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
 
@@ -12,12 +14,12 @@ public class OrdSettleRo {
 	/**
 	 * 已经结算的订单总额
 	 */
-	private Long AlreadySettle;
+	private BigDecimal AlreadySettle;
 	
 	/**
 	 * 没有结算
 	 */
-	private Long notSettle;
+	private BigDecimal notSettle;
 		
 	
 }
